new function - migration_tools.make_stub_bib()
Simple function to create a stub MARCXML bib from a set of columns.
The first argument is an array of tag/subfield specifiers, e.g., ARRAY['001', '245a', '500a'].
The second argument is an array of text containing the values to plugin into each field.
If the value for a given field is NULL or the empty string, it is not inserted.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>