eg_staged_bib_overlay: add --link-skipped
The new --link-skipped switch modifies the link_auth_bibs
action to say that bibs that were skipped due to having been
edited after the cutoff should be linked to authorities (rather
than linking imported bibs).

The use case for this: the library decides that they don't
want to scribble over the skipped bibs, but still want to
take advantage of any new authorities that correspond to them.

Signed-off-by: Galen Charlton <gmc@esilibrary.com>