Evergreen DB check for purged patrons with extant preferred name values
This is the inaugral test in what may turn into a set of pgTAP-based
tests for various data and schema problems that may occur in Evergreen
databases. This specific check looks for cases where a patron record
was purged but still has values in the preferred name and/or name
keywords columns due to bug LP#1802166.

To run all of the tests, go to db_checks/evergreen and run

pg_prove -v -U evergreen -r .

Signed-off-by: Galen Charlton <gmc@equinoxinitiative.org>