Add dbport option to eg_staged_bib_overlay
Allow the user to specify the port to use when connection to
PostgreSQL via a --dport command line option.  If the port is not
specified, the default PostgreSQL TCP port of 5432 is used.

This is useful for testing with different PostgreSQL installations on
the same server or for sites that my run PostgreSQL on a different
port for some reason.

Signed-off-by: Jason Stephenson <jason@sigio.com>